Update to current version of informalin (moved model definition to informalin project)
