+++ common/src/main/java/edu/kit/kastel/mcse/ardoco/core/api/data/Confidence.java
@@ -2,35 +2,34 @@
 package edu.kit.kastel.mcse.ardoco.core.api.data;
 
 import java.util.ArrayList;
-import java.util.Collection;
 import java.util.List;
 import java.util.Objects;
-import java.util.function.ToDoubleFunction;
 import java.util.stream.Collectors;
 
 import org.eclipse.collections.api.tuple.Pair;
 import org.eclipse.collections.impl.tuple.Tuples;
 
+import edu.kit.kastel.informalin.framework.common.AggregationFunctions;
 import edu.kit.kastel.informalin.framework.common.ICopyable;
 import edu.kit.kastel.mcse.ardoco.core.api.agent.IClaimant;
 
 public final class Confidence implements Comparable<Confidence>, ICopyable<Confidence> {
 
-    private final ConfidenceAggregator confidenceAggregator;
+    private final AggregationFunctions confidenceAggregator;
 
     private List<Pair<IClaimant, Double>> agentConfidence;
 
-    public Confidence(ConfidenceAggregator confidenceAggregator) {
+    public Confidence(AggregationFunctions confidenceAggregator) {
         this.confidenceAggregator = confidenceAggregator;
         this.agentConfidence = new ArrayList<>();
     }
 
-    public Confidence(IClaimant claimant, double probability, ConfidenceAggregator confidenceAggregator) {
+    public Confidence(IClaimant claimant, double probability, AggregationFunctions confidenceAggregator) {
         this(confidenceAggregator);
         this.addAgentConfidence(claimant, probability);
     }
 
-    private Confidence(ConfidenceAggregator confidenceAggregator, List<Pair<IClaimant, Double>> agentConfidence) {
+    private Confidence(AggregationFunctions confidenceAggregator, List<Pair<IClaimant, Double>> agentConfidence) {
         this(confidenceAggregator);
         this.agentConfidence = agentConfidence;
     }
@@ -58,11 +57,11 @@ public final class Confidence implements Comparable<Confidence>, ICopyable<Confi
         if (agentConfidence.isEmpty()) {
             return 0;
         }
-        if (confidenceAggregator == ConfidenceAggregator.ROLLING_AVERAGE) {
+        if (confidenceAggregator == AggregationFunctions.ROLLING_AVERAGE) {
             // No aggregate
             return confidenceAggregator.applyAsDouble(agentConfidence.stream().map(Pair::getTwo).toList());
         }
-        var groupAggregator = ConfidenceAggregator.MAX;
+        var groupAggregator = AggregationFunctions.MAX;
         var claimantGroupings = agentConfidence.stream().collect(Collectors.groupingBy(Pair::getOne)).values();
         var claimantConfidences = claimantGroupings.stream().map(l -> l.stream().map(Pair::getTwo).toList()).map(groupAggregator::applyAsDouble).toList();
         return confidenceAggregator.applyAsDouble(claimantConfidences);
@@ -77,7 +76,7 @@ public final class Confidence implements Comparable<Confidence>, ICopyable<Confi
      * @param localAggregator  aggregator for merging confidences of the same claimant
      * @return the combined confidence
      */
-    public static Confidence merge(Confidence a, Confidence b, ConfidenceAggregator globalAggregator, ConfidenceAggregator localAggregator) {
+    public static Confidence merge(Confidence a, Confidence b, AggregationFunctions globalAggregator, AggregationFunctions localAggregator) {
         var result = new Confidence(globalAggregator);
 
         for (var aConf : a.agentConfidence) {
@@ -115,70 +114,4 @@ public final class Confidence implements Comparable<Confidence>, ICopyable<Confi
         var other = (Confidence) obj;
         return Objects.equals(agentConfidence, other.agentConfidence) && confidenceAggregator == other.confidenceAggregator;
     }
-
-    public enum ConfidenceAggregator implements ToDoubleFunction<Collection<Double>> {
-        /**
-         * Use the median of the scores as final score.
-         */
-        MEDIAN(s -> {
-            var sortedNormalized = s.stream().sorted().toList();
-            var sizeHalf = sortedNormalized.size() / 2;
-
-            if (sortedNormalized.size() % 2 == 0) {
-                return (sortedNormalized.get(sizeHalf) + sortedNormalized.get(sizeHalf - 1)) / 2;
-            }
-            return sortedNormalized.get(sizeHalf);
-        }),
-
-        HARMONIC(s -> {
-            var quotient = s.stream().mapToDouble(d -> 1.0 / d).sum();
-            return s.size() / quotient;
-        }),
-
-        ROOTMEANSQUARED(s -> {
-            var squaredValuesSum = s.stream().mapToDouble(d -> Math.pow(d, 2)).sum();
-            return Math.sqrt(squaredValuesSum / s.size());
-        }),
-
-        CUBICMEAN(s -> {
-            var squaredValuesSum = s.stream().mapToDouble(d -> Math.pow(d, 3)).sum();
-            return Math.cbrt(squaredValuesSum / s.size());
-        }),
-
-        USE_MOST_RECENT(s -> s.stream().reduce((first, second) -> second).orElse(0.0)),
-
-        /**
-         * Use the average of the scores as final score.
-         */
-        AVERAGE(s -> s.stream().mapToDouble(d -> d).average().getAsDouble()),
-
-        ROLLING_AVERAGE(s -> s.stream().mapToDouble(d -> d).reduce((a, b) -> (a + b) / 2).getAsDouble()),
-
-        /**
-         * Use the max value of the scores as final score.
-         */
-        MAX(s -> s.stream().mapToDouble(d -> d).max().getAsDouble()),
-
-        /**
-         * Use the min value of the scores as final score.
-         */
-        MIN(s -> s.stream().mapToDouble(d -> d).min().getAsDouble()),
-
-        /**
-         * Use the sum of the scores as final score.
-         */
-        SUM(s -> s.stream().mapToDouble(d -> d).sum());
-
-        private final ToDoubleFunction<Collection<Double>> function;
-
-        ConfidenceAggregator(ToDoubleFunction<Collection<Double>> function) {
-            this.function = function;
-        }
-
-        @Override
-        public double applyAsDouble(Collection<Double> value) {
-            return this.function.applyAsDouble(value);
-        }
-    }
-
 }
