+++ model-provider/pom.xml
@@ -15,24 +15,16 @@
 
   <dependencies>
     <dependency>
-      <groupId>org.fuchss</groupId>
-      <artifactId>xml-object-mapper</artifactId>
-      <version>0.3</version>
-    </dependency>
-    <dependency>
       <groupId>com.fasterxml.jackson.core</groupId>
       <artifactId>jackson-databind</artifactId>
-      <version>${jackson.version}</version>
     </dependency>
-    <dependency>
-      <groupId>edu.kit.kastel.mcse.ardoco</groupId>
-      <artifactId>java-model</artifactId>
-      <version>0.1-SNAPSHOT</version>
-    </dependency>
-
     <dependency>
       <groupId>edu.kit.kastel.mcse.ardoco</groupId>
       <artifactId>common</artifactId>
+    </dependency>
+    <dependency>
+      <groupId>edu.kit.kastel.informalin</groupId>
+      <artifactId>models</artifactId>
     </dependency>
 
     <dependency>
@@ -41,13 +33,4 @@
       <scope>test</scope>
     </dependency>
   </dependencies>
-
-  <repositories>
-    <!-- TODO Remove as soon as published to Maven Central -->
-    <repository>
-      <id>gh-fuchss</id>
-      <name>GitHub Maven Packages</name>
-      <url>https://packages.fuchss.org/github/releases/raw/branch/releases/</url>
-    </repository>
-  </repositories>
 </project>
