+++ model-provider/src/main/java/edu/kit/kastel/mcse/ardoco/core/model/ModelProvider.java
@@ -1,5 +1,5 @@
 /* Licensed under MIT 2021-2022. */
-package edu.kit.kastel.mcse.ardoco.core.model.provider;
+package edu.kit.kastel.mcse.ardoco.core.model;
 
 import java.util.Map;
 
@@ -8,8 +8,6 @@ import org.eclipse.collections.api.list.ImmutableList;
 import edu.kit.kastel.mcse.ardoco.core.api.data.model.IModelInstance;
 import edu.kit.kastel.mcse.ardoco.core.api.data.model.IModelRelation;
 import edu.kit.kastel.mcse.ardoco.core.api.data.model.IModelState;
-import edu.kit.kastel.mcse.ardoco.core.model.IModelConnector;
-import edu.kit.kastel.mcse.ardoco.core.model.ModelExtractionState;
 
 /**
  * The model extractor extracts the instances and relations via an connector. The extracted items are stored in a model
