+++ model-provider/src/test/java/edu/kit/kastel/mcse/ardoco/core/model/java/JavaJsonModelConnectorTest.java
@@ -10,6 +10,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import edu.kit.kastel.mcse.ardoco.core.api.data.model.Metamodel;
+import edu.kit.kastel.mcse.ardoco.core.model.JavaJsonModelConnector;
 
 class JavaJsonModelConnectorTest {
     @Test
