+++ model-provider/src/test/java/edu/kit/kastel/mcse/ardoco/core/model/pcm/PcmXMLModelConnectorTest.java
@@ -19,6 +19,7 @@ import org.xml.sax.SAXException;
 
 import edu.kit.kastel.mcse.ardoco.core.api.data.model.IModelInstance;
 import edu.kit.kastel.mcse.ardoco.core.api.data.model.Metamodel;
+import edu.kit.kastel.mcse.ardoco.core.model.PcmXMLModelConnector;
 
 class PcmXMLModelConnectorTest {
     private static final Logger logger = LogManager.getLogger();
