+++ tests/src/test/java/edu/kit/kastel/mcse/ardoco/core/tests/inconsistencies/eval/text/TextRandomizer.java
@@ -1,19 +1,10 @@
-/* Licensed under MIT 2021. */
+/* Licensed under MIT 2021-2022. */
 package edu.kit.kastel.mcse.ardoco.core.tests.inconsistencies.eval.text;
 
-import java.io.BufferedReader;
-import java.io.File;
-import java.io.FileNotFoundException;
-import java.io.FileReader;
-import java.io.FileWriter;
-import java.io.IOException;
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.Random;
+import java.io.*;
+import java.util.*;
 
-import edu.kit.kastel.mcse.ardoco.core.tests.inconsistencies.misc.Pair;
+import edu.kit.kastel.informalin.framework.common.tuple.Pair;
 
 public class TextRandomizer {
