--- tests/src/test/java/edu/kit/kastel/mcse/ardoco/core/tests/inconsistencies/misc/Pair.java
@@ -1,21 +1,0 @@
-/* Licensed under MIT 2021. */
-package edu.kit.kastel.mcse.ardoco.core.tests.inconsistencies.misc;
-
-public class Pair<L, R> {
-    private L left;
-    private R right;
-
-    public Pair(L left, R right) {
-        this.left = left;
-        this.right = right;
-    }
-
-    public L getLeft() {
-        return left;
-    }
-
-    public R getRight() {
-        return right;
-    }
-
-}
